Automatically add Haunt's modules to Guile's load path at startup.
This fixes a longstanding usability issue where people would install
Haunt to /usr/local and then wonder why Haunt doesn't work, which was
because Guile's default load path is /usr and thus could not find the
newly installed modules in /usr/local.  I received several emails
about this over the years and didn't know to deal with it.  Many
thanks to Daniel Meißner for pointing out that Guix had already found
a solution to this problem because of course Guix would have this
problem solved!  Ludovic is an autotools wizard.